Move AnalyticsManger remote data store related classes to the analytics module
COSDK-500
