+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/DefaultAnalyticsRemoteDataStore.kt
@@ -8,10 +8,10 @@
 
 package com.adyen.checkout.core.analytics.internal.data.remote
 
-import com.adyen.checkout.core.internal.data.api.AnalyticsService
-import com.adyen.checkout.core.internal.data.model.AnalyticsSetupRequest
-import com.adyen.checkout.core.internal.data.model.AnalyticsSetupResponse
-import com.adyen.checkout.core.internal.data.model.AnalyticsTrackRequest
+import com.adyen.checkout.core.analytics.internal.data.remote.api.AnalyticsService
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupResponse
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackRequest
 
 internal class DefaultAnalyticsRemoteDataStore(
     private val analyticsService: AnalyticsService,
