+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/DefaultAnalyticsSetupProvider.kt
@@ -13,7 +13,7 @@ import android.os.Build
 import com.adyen.checkout.core.data.model.Amount
 import com.adyen.checkout.core.analytics.internal.AnalyticsPlatformParams
 import com.adyen.checkout.core.analytics.internal.AnalyticsSource
-import com.adyen.checkout.core.internal.data.model.AnalyticsSetupRequest
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
 import com.adyen.checkout.core.analytics.internal.AnalyticsParamsLevel
 import java.util.Locale
