+++ core/src/test/java/com/adyen/checkout/core/internal/analytics/data/remote/AnalyticsTrackRequestProviderTest.kt
@@ -3,10 +3,10 @@ package com.adyen.checkout.core.internal.analytics.data.remote
 import com.adyen.checkout.core.analytics.internal.AnalyticsEvent
 import com.adyen.checkout.core.analytics.internal.DirectAnalyticsEventCreation
 import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsTrackRequestProvider
-import com.adyen.checkout.core.internal.data.model.AnalyticsTrackError
-import com.adyen.checkout.core.internal.data.model.AnalyticsTrackInfo
-import com.adyen.checkout.core.internal.data.model.AnalyticsTrackLog
-import com.adyen.checkout.core.internal.data.model.AnalyticsTrackRequest
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackError
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackInfo
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackLog
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackRequest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
