+++ core/src/test/java/com/adyen/checkout/core/internal/analytics/data/remote/DefaultAnalyticsSetupProviderTest.kt
@@ -7,7 +7,7 @@ import android.util.DisplayMetrics
 import com.adyen.checkout.core.data.model.Amount
 import com.adyen.checkout.core.analytics.internal.AnalyticsPlatformParams
 import com.adyen.checkout.core.analytics.internal.AnalyticsSource
-import com.adyen.checkout.core.internal.data.model.AnalyticsSetupRequest
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
 import com.adyen.checkout.core.analytics.internal.AnalyticsParamsLevel
 import com.adyen.checkout.core.analytics.internal.data.remote.DefaultAnalyticsSetupProvider
 import org.junit.jupiter.api.Assertions.assertEquals
