+++ core/src/test/java/com/adyen/checkout/core/internal/data/api/AnalyticsServiceTest.kt
@@ -1,6 +1,7 @@
 package com.adyen.checkout.core.internal.data.api
 
-import com.adyen.checkout.core.internal.data.model.AnalyticsTrackRequest
+import com.adyen.checkout.core.analytics.internal.data.remote.api.AnalyticsService
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackRequest
 import com.adyen.checkout.core.internal.data.model.EmptyResponse
 import com.adyen.checkout.test.LoggingExtension
 import kotlinx.coroutines.test.runTest
