+++ core/src/test/java/com/adyen/checkout/core/internal/data/model/AnalyticsSetupResponseTest.kt
@@ -1,5 +1,6 @@
 package com.adyen.checkout.core.internal.data.model
 
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupResponse
 import org.json.JSONObject
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
