+++ core/src/test/java/com/adyen/checkout/core/internal/data/model/AnalyticsTrackRequestTest.kt
@@ -1,5 +1,9 @@
 package com.adyen.checkout.core.internal.data.model
 
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackError
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackInfo
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackLog
+import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackRequest
 import org.json.JSONObject
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
